Fix issues with Camel Drools6 Example to deploy it as a bundle or using profile on JBoss Fuse, rename package to use io.fabric8, add some missing dependencies in the BOM
