+++ tooling/examples/camel-drools-example/src/main/fabric8/io.fabric8.drools.example.properties
@@ -1,0 +1,10 @@
+org.ops4j.pax.url.mvn.repositories= \
+    http://repo1.maven.org/maven2@id=central, \
+    https://repo.fusesource.com/nexus/content/groups/public@id=fusepublic, \
+    https://repository.jboss.org/nexus/content/repositories/public@id=jbosspublic, \
+    https://repo.fusesource.com/nexus/content/repositories/releases@id=jbossreleases, \
+    https://repo.fusesource.com/nexus/content/groups/ea@id=jbossearlyaccess, \
+    http://repository.springsource.com/maven/bundles/release@id=ebrreleases, \
+    http://repository.springsource.com/maven/bundles/external@id=ebrexternal, \
+    http://oss.sonatype.org/content/groups/scala-tools@id=scala, \
+    https://repository.jboss.org/nexus/content/repositories/snapshots@snapshots
