+++ tooling/examples/camel-drools-example/src/main/fabric8/org.ops4j.pax.url.mvn.properties
@@ -1,0 +1,1 @@
+org.ops4j.pax.url.mvn.repositories = ${profile:io.fabric8.drools.example/org.ops4j.pax.url.mvn.repositories},file:${karaf.data}/maven/agent@snapshots@id=agent-local
