+++ tooling/examples/camel-drools-example/src/main/resources/META-INF/spring/camel-context.xml
@@ -5,78 +5,84 @@
        xsi:schemaLocation="
        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
        http://camel.apache.org/schema/spring       http://camel.apache.org/schema/spring/camel-spring.xsd
-       http://drools.org/schema/kie-spring         org/kie/spring/kie-spring-2.0.0.xsd">
+       http://drools.org/schema/kie-spring http://drools.org/schema/kie-spring.xsd">
 
-  <!-- Beans used by camel in the example -->
-  <bean id="droolsHelper" class="org.fusesource.example.drools.DroolsCommandHelper"/>
-  <bean id="personHelper" class="org.fusesource.example.drools.PersonHelper"/>
+    <!-- Beans used by camel in the example -->
+    <bean id="droolsHelper" class="io.fabric8.example.drools.DroolsCommandHelper"/>
+    <bean id="personHelper" class="io.fabric8.example.drools.PersonHelper"/>
 
-  <kie:ksession-ref id="ksession1">
-    <kie:batch>
-      <kie:set-global identifier="list" >
-        <bean class="java.util.ArrayList" />
-      </kie:set-global>
-    </kie:batch>
-  </kie:ksession-ref>
+    <bean id="kiePostProcessor" class="org.drools.osgi.spring.OsgiKModuleBeanFactoryPostProcessor"/>
 
-  <!-- Camel Route -->
-  <camelContext xmlns="http://camel.apache.org/schema/spring">
+    <kie:kmodule id="camel-context">
+        <kie:kbase name="drl">
+            <kie:ksession name="ksession1" type="stateless">
+                <kie:batch>
+                    <kie:set-global identifier="list" >
+                        <bean class="java.util.ArrayList" />
+                    </kie:set-global>
+                </kie:batch>
+            </kie:ksession>
+        </kie:kbase>
+    </kie:kmodule>
 
-    <route trace="false" id="testRoute">
-      <description>Example route that will regularly create a Person with a random age and verify their age
-      </description>
-      <from uri="timer:testRoute?period=10s"/>
-      <bean method="createTestPerson" ref="personHelper"/>
-      <to uri="kie:ksession1?action=insertBody" id="AgeVerification">
-        <description>valid 'action' values are:
-          'execute' that takes a 'Command' object (default)
-          'insertBody' that inserts the Exchange.in.body, and executes rules
-          'insertMessage' that inserts the Exchange.in (type org.apache.camel.Message), and executes rules
-          'insertExchange' that inserts the Exchange (type org.apache.camel.Exchange), and executes rules
-        </description>
-      </to>
-      <choice>
-        <when id="CanDrink">
-          <simple>${body.canDrink}</simple>
-          <log logName="Bar" message="Person ${body.name} can go to the bar"/>
-        </when>
-        <otherwise>
-          <log logName="Home" message="Person ${body.name} is staying home"/>
-        </otherwise>
-      </choice>
-    </route>
+    <!-- Camel Route -->
+    <camelContext xmlns="http://camel.apache.org/schema/spring">
 
-    <route trace="false" id="ruleOnBody">
-      <description>This route shows an example of passing (inserting) the Body of the message as a POJO to Drools
-        for execution
-      </description>
-      <from uri="direct://ruleOnBody"/>
-      <to uri="kie:ksession1?action=insertBody">
-        <description>valid 'action' values are:
-          'execute' that takes a 'Command' object (default)
-          'insertBody' that inserts the Exchange.in.body, and executes rules
-          'insertMessage' that inserts the Exchange.in (type org.apache.camel.Message), and executes rules
-          'insertExchange' that inserts the Exchange (type org.apache.camel.Exchange), and executes rules
-        </description>
-      </to>
-    </route>
+        <route trace="false" id="testRoute">
+            <description>Example route that will regularly create a Person with a random age and verify their age
+            </description>
+            <from uri="timer:testRoute?period=10s"/>
+            <bean method="createTestPerson" ref="personHelper"/>
+            <to uri="kie:ksession1?action=insertBody" id="AgeVerification">
+                <description>valid 'action' values are:
+                    'execute' that takes a 'Command' object (default)
+                    'insertBody' that inserts the Exchange.in.body, and executes rules
+                    'insertMessage' that inserts the Exchange.in (type org.apache.camel.Message), and executes rules
+                    'insertExchange' that inserts the Exchange (type org.apache.camel.Exchange), and executes rules
+                </description>
+            </to>
+            <choice>
+                <when id="CanDrink">
+                    <simple>${body.canDrink}</simple>
+                    <log logName="Bar" message="Person ${body.name} can go to the bar"/>
+                </when>
+                <otherwise>
+                    <log logName="Home" message="Person ${body.name} is staying home"/>
+                </otherwise>
+            </choice>
+        </route>
 
-    <route trace="false" id="ruleOnCommand">
-      <description>This route shows an example of creating a Drools Command wrapping the Body of the message with
-        the help of the custom DroolsCommandHelperBean (included in this project), and passing this Drools
-        Command to the ruleset for execution.
-      </description>
-      <from uri="direct://ruleOnCommand"/>
-      <bean method="insertAndFireAll" ref="droolsHelper"/>
-      <to uri="kie:ksession1?action=execute">
-        <description>valid 'action' values are:
-          'execute' that takes a 'Command' object (default)
-          'insertBody' that inserts the Exchange.in.body, and executes rules
-          'insertMessage' that inserts the Exchange.in (type org.apache.camel.Message), and executes rules
-          'insertExchange' that inserts the Exchange (type org.apache.camel.Exchange), and executes rules
-        </description>
-      </to>
-    </route>
-  </camelContext>
+        <route trace="false" id="ruleOnBody">
+            <description>This route shows an example of passing (inserting) the Body of the message as a POJO to Drools
+                for execution
+            </description>
+            <from uri="direct://ruleOnBody"/>
+            <to uri="kie:ksession1?action=insertBody">
+                <description>valid 'action' values are:
+                    'execute' that takes a 'Command' object (default)
+                    'insertBody' that inserts the Exchange.in.body, and executes rules
+                    'insertMessage' that inserts the Exchange.in (type org.apache.camel.Message), and executes rules
+                    'insertExchange' that inserts the Exchange (type org.apache.camel.Exchange), and executes rules
+                </description>
+            </to>
+        </route>
+
+        <route trace="false" id="ruleOnCommand">
+            <description>This route shows an example of creating a Drools Command wrapping the Body of the message with
+                the help of the custom DroolsCommandHelperBean (included in this project), and passing this Drools
+                Command to the ruleset for execution.
+            </description>
+            <from uri="direct://ruleOnCommand"/>
+            <bean method="insertAndFireAll" ref="droolsHelper"/>
+            <to uri="kie:ksession1?action=execute">
+                <description>valid 'action' values are:
+                    'execute' that takes a 'Command' object (default)
+                    'insertBody' that inserts the Exchange.in.body, and executes rules
+                    'insertMessage' that inserts the Exchange.in (type org.apache.camel.Message), and executes rules
+                    'insertExchange' that inserts the Exchange (type org.apache.camel.Exchange), and executes rules
+                </description>
+            </to>
+        </route>
+    </camelContext>
 
 </beans>
