+++ tooling/examples/camel-drools-example/src/test/resources/decision/IntegrationExampleTest.xls
