+++ tooling/examples/camel-drools-example/src/test/resources/drl/can-drink-rule.drl
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012 Red Hat
+ * Copyright 2014 Red Hat
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -11,27 +11,15 @@
  *    distributed under the License is distributed on an "AS IS" BASIS,
  *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  *    See the License for the specific language governing permissions and
- *    limitations under the License.
+ *    limitations under the Licene.
  */
+package io.fabric8.example.drools;
 
-package org.fusesource.example.drools;
+import io.fabric8.example.drools.Person;
 
-import java.util.Random;
-
-public class CheeseHelper {
-
-    private final Random random = new Random();
-
-    public Cheese createTestCheese() {
-        Cheese cheese = new Cheese();
-        if (random.nextBoolean()) {
-            cheese.setType("Stilton");
-            cheese.setPrice(10);
-        } else {
-            cheese.setType("Cheddar");
-            cheese.setPrice(50);
-        }
-        return cheese;
-    }
-
-}
+rule "CanDrink"
+when
+    p : Person( age >= 21 )
+then
+	p.setCanDrink(true);
+end
